Create WorkerView To fix: null pointer exception on fetching user Works: logout redirecting to login

+++ src/main/java/com/example/application/data/service/ParkingSlotRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ParkingSlotRepository extends JpaRepository<ParkingSlot, Integer> {
+public interface ParkingSlotRepository extends JpaRepository<ParkingSlot, Long> {
 
     List<ParkingSlot> getByName(String name);
 }
